chore(move):미션4파일 정리
